app_streamlit/requirements.txt
streamlit
pandas
openpyxl
google-generativeai
python-dotenv
requests-oauthlib
openai
lloom
git+https://github.com/michelle123lam/lloom.git
pyyaml
bertopic
sentence-transformers